flask-api/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai
gunicorn==21.2.0
